Added right click mode and moved PixiEditorSettings.cs into .CommonApi.UserPreferences.Settings namespace
