+++ src/PixiEditor.AvaloniaUI/Models/ExceptionHandling/CrashReport.cs
@@ -15,6 +15,7 @@ using PixiEditor.AvaloniaUI.ViewModels.Document;
 using PixiEditor.AvaloniaUI.Views;
 using PixiEditor.Extensions.Common.Localization;
 using PixiEditor.Extensions.CommonApi.UserPreferences;
+using PixiEditor.Extensions.CommonApi.UserPreferences.Settings.PixiEditor;
 using PixiEditor.Parser;
 
 namespace PixiEditor.AvaloniaUI.Models.ExceptionHandling;
