+++ src/PixiEditor.AvaloniaUI/Models/Services/NewsFeed/NewsProvider.cs
@@ -5,6 +5,7 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 using PixiEditor.Extensions.CommonApi.UserPreferences.Settings;
+using PixiEditor.Extensions.CommonApi.UserPreferences.Settings.PixiEditor;
 using PixiEditor.Platform;
 
 namespace PixiEditor.AvaloniaUI.Models.Services.NewsFeed;
