+++ src/PixiEditor.AvaloniaUI/ViewModels/SubViewModels/StylusViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using PixiEditor.AvaloniaUI.Models.Commands.Attributes.Commands;
 using PixiEditor.AvaloniaUI.ViewModels.Tools;
 using PixiEditor.AvaloniaUI.ViewModels.Tools.Tools;
-using PixiEditor.Extensions.CommonApi.UserPreferences.Settings;
+using PixiEditor.Extensions.CommonApi.UserPreferences.Settings.PixiEditor;
 
 namespace PixiEditor.AvaloniaUI.ViewModels.SubViewModels;
